refactor(cosh-ng): [shell] split raw_relay emit
The watchdog addition pushed raw_relay.rs to 717 production lines,
past the 700-line layout gate. Extract the PTY handoff emission
domain (resolve_pty_emit, emit_to_pty,
restore_prompt_display_before_handoff, write_handoff_request) into
the raw_relay/pty_emit.rs owner submodule, following the existing
input_events/terminal_recovery/terminal_size submodule pattern from
main. Function bodies are moved verbatim; behavior is unchanged and
crates/cosh-shell/scripts/check-layout.sh passes again.
